Chore: Moved Global Redis task to ws.rs
