+++ src/routes/ws.rs
@@ -17,6 +17,7 @@ use serde::{Deserialize, Serialize};
 use serde_json;
 
 use redis::AsyncCommands;
+use tokio::sync::broadcast;
 use tokio::sync::mpsc::{Receiver, Sender};
 use tokio::task::JoinHandle;
 
@@ -191,6 +192,51 @@ pub async fn ws_handler(
     ws.on_upgrade(move |socket| WsConnection::new(state, socket))
 }
 
-pub fn routes() -> Router<SharedState> {
+fn spawn_global_redis_listener_task(
+    mut redis: redis::aio::ConnectionManager,
+    redis_client: redis::Client,
+    redis_task_tx: broadcast::Sender<Timer>,
+) -> JoinHandle<()> {
+    tokio::spawn(async move {
+        let mut connection = redis_client.get_async_connection().await.unwrap();
+        let _: () = redis::cmd("CONFIG")
+            .arg("SET")
+            .arg("notify-keyspace-events")
+            .arg("KEA")
+            .query_async(&mut connection)
+            .await
+            .unwrap();
+
+        let mut pubsub = connection.into_pubsub();
+
+        pubsub
+            .psubscribe("__keyspace@0__:*")
+            .await
+            .expect("Failed to subscribe to redis channel");
+
+        let mut pubsub = pubsub.into_on_message();
+
+        while let Some(msg) = pubsub.next().await {
+            println!("Updated! {:?}", msg);
+            let timer_id = msg.get_channel_name().split(":").last().unwrap();
+
+            let timer_str = &redis
+                .get::<String, String>(String::from(timer_id))
+                .await
+                .expect("Did not find timer in redis");
+            let timer: Timer = serde_json::from_str(timer_str).unwrap();
+
+            // Broadcast to all listeners
+            redis_task_tx.send(timer).unwrap();
+        }
+    })
+}
+
+pub fn routes(
+    redis: redis::aio::ConnectionManager,
+    redis_client: redis::Client,
+    redis_task_tx: broadcast::Sender<Timer>,
+) -> Router<SharedState> {
+    spawn_global_redis_listener_task(redis, redis_client, redis_task_tx);
     Router::new().route("/", get(ws_handler))
 }
